v0.8 Sprint 1 (QA-013b): extract NativeTcpReactor from NativeTcpCarrier — close CognitiveComplexity
Second of the multi-PR QA-013 decomposition (after QA-013a socket-
backend extract). Lifts the reactor selector loop out of the carrier
as a top-level package-private class — same threading model and
single-MPSC-queue / wakeup coalescing semantics, but now ownership
is explicit rather than buried as a non-static inner class.

What moves:

  NativeTcpCarrier.ReactorLoop         → NativeTcpReactor (top-level)
  NativeTcpCarrier.ReactorRequest      → NativeTcpReactor.ReactorRequest
  NativeTcpCarrier.ReactorRequestType  → NativeTcpReactor.ReactorRequestType

How the host coupling is preserved:

  - NativeTcpReactor holds a final NativeTcpCarrier host field, set
    in the constructor (this passed by the carrier in
    startServerRuntime).
  - All callbacks (readIngress, flushStream, closeKeyStream,
    handleAsyncFailure, resolveStream) now go through host.xxx().
    Carrier's visibility was upgraded from private to package-private
    on those five methods (no public surface change).
  - A new package-private getter isRunning() proxies running.get().
  - The two on-error fallbacks that previously read streamByChannel
    directly (registerChannel IOException branch) now route through
    host.resolveStream(channel) — same semantics
    (ChannelRuntimeRegistry.resolveStream already falls back to
    streamByChannel internally).
  - ChannelRuntimeRegistry references update verbatim:
    NativeTcpCarrier.ReactorLoop → NativeTcpReactor (3 fields + 1
    state record + 2 method signatures).

Suppression delta on NativeTcpCarrier (incremental):

  removed: PMD.CognitiveComplexity  (1 closed — runLoop now lives in
                                      NativeTcpReactor under its own
                                      method-level suppression with
                                      rationale)
  retained: 10 of the original 11 (TooManyMethods + ExcessiveImports
            + CouplingBetweenObjects + GodClass + CyclomaticComplexity
            + ExceptionAsFlowControl + AvoidCatchingGenericException
            + CloseResource + LawOfDemeter +
            AvoidInstantiatingObjectsInLoops)

Suppressions on the new NativeTcpReactor (all with inline rationale):

  class-level: AvoidCatchingGenericException + CloseResource +
               CyclomaticComplexity + TooManyMethods
  runLoop():   CognitiveComplexity + CyclomaticComplexity (selector
               loop drains MPSC, polls keys, dispatches READ/WRITE —
               flat structure)

Delta:
  NativeTcpCarrier 1006 → 803 LOC (-203 / -20%)
  + NativeTcpReactor 259 LOC
  ChannelRuntimeRegistry — 14 line-touches (type rename only)
  Class-level suppressions on carrier: 11 → 10 (-1)

Cumulative QA-013 progress so far (QA-013a + QA-013b):
  NativeTcpCarrier 1381 → 803 LOC (-578 / -42%)
  Class-level suppressions on carrier: 11 → 10 (-1)

QA-013c (client ingress/writer pump extract) will likely close
AvoidInstantiatingObjectsInLoops and possibly
AvoidCatchingGenericException once the pump bodies move out.

Test plan:
  - Full transport test suite 110/110 green (CarrierPinningTckTest,
    MultiReactor{2,4}TckTest, NativeTcpCarrierTckTest, NativeTcpStream
    TckTest, IngressIntegrationTest, TransportStressTest — including
    TLS / plaintext / FFM-armed / NIO-fallback paths)
  - Full reactor verify SUCCESS (SPI + TCK + Core + Community Testkit
    + Community + Build Config + Parent + Root)
  - PMD 0 violations on community module
  - Checkstyle 0 violations on community module
  - Local JDK 26+35 G1 crash worked around with -XX:+UseSerialGC
    (pre-existing memory note; CI uses VT scheduler bump + drain
    budget — unrelated to this refactor)
